Moved Grid.setColumnOrder() tests to a separate test class
